+++ build.gradle
@@ -110,6 +110,7 @@ dependencies {
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
     implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
     implementation 'org.springframework.boot:spring-boot-starter-mail'
+    implementation 'org.apache.commons:commons-math3:3.6.1'
 
     implementation 'org.springframework.boot:spring-boot-starter-validation'
